# contracts/openapi/gateway-node.openapi.yaml
openapi: 3.1.0
info:
  title: Ishtirak gateway (public BFF)
  version: 0.1.0
  description: >
    The public API the web client consumes (ADR-003). The gateway authenticates
    every request (Bearer JWT), derives operatorId + role from the token, and
    aggregates/proxies to the internal core-java and analytics-python services.
    All resources are implicitly scoped to the caller's operator — clients never
    pass operatorId.


    The real-time channel (`GET /api/ws`, WebSocket upgrade) is **not** modeled
    in OpenAPI; its protocol lives in ../asyncapi.yaml and ../../docs/API.md.


    **Design-ahead-of-implementation:** only `/health` and `/ready` are built
    today (`x-status: built`); everything else is `x-status: planned`.
  contact: { name: Ishtirak, url: https://ishtirak.dev }
servers:
  - url: http://localhost:8080
    description: Local / docker-compose (public)
security:
  - bearerAuth: []
...
paths:
...
  /api/subscribers:
    get:
      tags: [subscribers]
      operationId: apiListSubscribers
      summary: List subscribers (paginated)
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      parameters:
        - $ref: 'components.yaml#/components/parameters/pageParam'
        - $ref: 'components.yaml#/components/parameters/limitParam'
      responses:
        '200':
          description: Page of subscribers.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: 'components.yaml#/components/schemas/Subscriber' }
                  meta: { $ref: 'components.yaml#/components/schemas/PageMeta' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
    post:
      tags: [subscribers]
      operationId: apiCreateSubscriber
      summary: Create a subscriber
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, tierId]
              properties:
                name: { type: string }
                tierId: { type: string, format: uuid }
                meterId: { type: string }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Subscriber' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
  /api/subscribers/{id}:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
    get:
      tags: [subscribers]
      operationId: apiGetSubscriber
      summary: Get a subscriber
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      responses:
        '200':
          description: Subscriber.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Subscriber' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
    patch:
      tags: [subscribers]
      operationId: apiUpdateSubscriber
      summary: Update a subscriber
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                name: { type: string }
                tierId: { type: string, format: uuid }
                status: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        '200':
          description: Updated subscriber.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Subscriber' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/subscribers/{id}/readings:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
      - $ref: 'components.yaml#/components/parameters/pageParam'
      - $ref: 'components.yaml#/components/parameters/limitParam'
    get:
      tags: [readings]
      operationId: apiListSubscriberReadings
      summary: Reading history for a subscriber
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      responses:
        '200':
          description: Page of readings.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: 'components.yaml#/components/schemas/Reading' }
                  meta: { $ref: 'components.yaml#/components/schemas/PageMeta' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/tiers:
    get:
      tags: [tiers]
      operationId: apiListTiers
      summary: List amperage tiers
      x-status: planned
      responses:
        '200':
          description: Tiers.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: 'components.yaml#/components/schemas/Tier' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
    post:
      tags: [tiers]
      operationId: apiCreateTier
      summary: Create an amperage tier and tariff override
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, amperage, standingFeeUsd, standingFeeLbp, perKwhRateUsd, perKwhRateLbp]
              properties:
                name: { type: string }
                amperage: { type: integer, minimum: 1 }
                tariffPolicyOverride:
                  oneOf:
                    - { $ref: 'components.yaml#/components/schemas/TariffPolicy' }
                    - { type: 'null' }
                standingFeeUsd: { type: number, minimum: 0 }
                standingFeeLbp: { type: integer, minimum: 0 }
                perKwhRateUsd: { type: number, minimum: 0 }
                perKwhRateLbp: { type: integer, minimum: 0 }
      responses:
        '201':
          description: Created tier.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Tier' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
  /api/tiers/{id}:
    parameters:
      - $ref: 'components.yaml#/components/parameters/idParam'
    get:
      tags: [tiers]
      operationId: apiGetTier
      summary: Get a tier
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      responses:
        '200':
          description: Tier.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Tier' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
    patch:
      tags: [tiers]
      operationId: apiUpdateTier
      summary: Update tier pricing, policy override, or status
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                name: { type: string }
                amperage: { type: integer, minimum: 1 }
                tariffPolicyOverride:
                  oneOf:
                    - { $ref: 'components.yaml#/components/schemas/TariffPolicy' }
                    - { type: 'null' }
                standingFeeUsd: { type: number, minimum: 0 }
                standingFeeLbp: { type: integer, minimum: 0 }
                perKwhRateUsd: { type: number, minimum: 0 }
                perKwhRateLbp: { type: integer, minimum: 0 }
                status: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        '200':
          description: Updated tier.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Tier' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
        '404': { $ref: 'components.yaml#/components/responses/NotFound' }
  /api/readings:
    post:
      tags: [readings]
      operationId: apiRecordReading
      summary: Record a meter reading
      description: Proxies to core-java, which emits `reading.recorded`.
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN, OPERATOR_STAFF]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [subscriberId, kwh, readingAt]
              properties:
                subscriberId: { type: string, format: uuid }
                kwh: { type: number, minimum: 0 }
                readingAt: { type: string, format: date-time }
      responses:
        '201':
          description: Recorded reading.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Reading' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
...
  /api/me/readings:
    get:
      tags: [readings]
      operationId: apiListMyReadings
      summary: The caller's own readings (subscriber portal)
      x-status: planned
      x-required-roles: [SUBSCRIBER]
      parameters:
        - $ref: 'components.yaml#/components/parameters/pageParam'
        - $ref: 'components.yaml#/components/parameters/limitParam'
      responses:
        '200':
          description: Page of the subscriber's readings.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: 'components.yaml#/components/schemas/Reading' }
                  meta: { $ref: 'components.yaml#/components/schemas/PageMeta' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
...
  /api/outages:
    get:
      tags: [outages]
      operationId: apiListOutages
      summary: Scheduled / active outages (operator or subscriber view)
      x-status: planned
      responses:
        '200':
          description: Outages.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: 'components.yaml#/components/schemas/Outage' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
    post:
      tags: [outages]
      operationId: apiScheduleOutage
      summary: Schedule a load-shedding window
      description: Proxies to core-java, which emits `outage.scheduled`.
      x-status: planned
      x-required-roles: [OPERATOR_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [startsAt, endsAt, reason]
              properties:
                startsAt: { type: string, format: date-time }
                endsAt: { type: string, format: date-time }
                reason: { type: string, enum: [FUEL, MAINTENANCE, GRID, OTHER] }
      responses:
        '201':
          description: Scheduled outage.
          content:
            application/json:
              schema: { $ref: 'components.yaml#/components/schemas/Outage' }
        '400': { $ref: 'components.yaml#/components/responses/BadRequest' }
        '401': { $ref: 'components.yaml#/components/responses/Unauthorized' }
        '403': { $ref: 'components.yaml#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      $ref: 'components.yaml#/components/securitySchemes/bearerAuth'